requirements.txt
peft
transformers
PyYAML
datasets
nltk
torch
pandas

# Metrics
bert_score
sentence-transformers

# Gridsearch
optuna

# plotting 
matplotlib
plotly
kaleido

# Development
black
isort